Cache the reference counted VLC objects and their event managers
We need to avoid garbage collecting a python function if it is attached to
an event manager that could still send an event. However, to avoid leaking
memory we still need to eventually delete the python functions and related state.
There's no event telling you when a libvlc object is destroyed, so to safely
delete the state on the python side the python code must track the object's
reference count and assume the object is immediately destroyed when the count
drops to zero.

This change adds a map caching all VLC objects that are explicitly or
implicitly retained on the python side. Each cached object includes a
reference count and when that count drops to zero all of the event listeners
are detached for that object. The underlying VLC object may still be
referenced inside libvlc and may not be actually destroyed, so don't release
the object if you want to continue receiving events from it.